% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.ReadORFasta.R
\name{Function.ReadORFasta}
\alias{Function.ReadORFasta}
\title{Function.ReadORFasta: function to read and format a fasta file containing multiple OR amino acid sequences}
\usage{
Function.ReadORFasta(string.PathFastaFile)
}
\arguments{
\item{string.PathFastaFile}{String containing system path to fasta file. Make sure: no slash characters in the sequence names}
}
\value{
dataframe.ORSequences Returns a data.frame with four columns: Species (genome), Sequence names (fasta names), sequence scaffolds (taken from fasta file), and sequence strings (amino acid strings for each OR gene)
}
\description{
This function returns a data frame with the first column representing the OR sequence names, and the second column containing a string for each OR amino acid sequence
}
\examples{

################
# load depends #
################
library(seqinr)

#########################
# read input fasta file #
#########################
handle.InputFastaFile <- Function.ReadORFasta(string.PathFastaFile = '~/Desktop/Beetle ORs Protein Translations.fasta')
}
\keyword{beetle}
\keyword{chemosensation,}
\keyword{coleoptera,}
\keyword{genomes,}
\keyword{olfactory}
\keyword{receptors,}
