% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.ReadResults_DeepTMHMM.R
\name{Function.ReadResults_DeepTMHMM}
\alias{Function.ReadResults_DeepTMHMM}
\title{Function.ReadResults_DeepTMHMM: function to read the results from DeepTMTHMM}
\usage{
Function.ReadResults_DeepTMHMM(
  string.Path_RESULTS,
  string.SequenceName,
  string.SpeciesName,
  string.SequenceScaffold
)
}
\arguments{
\item{string.Path_RESULTS}{String containing path to DeepTMHMM results}

\item{string.SequenceName}{String containing the name of the sequence (used to name the directory as well)}

\item{string.SpeciesName}{String containing the name of species (genome) taken from the original fasta file}

\item{string.SequenceScaffold}{String containing the name of the scaffold with putative BLAST hits}
}
\value{
vector.RESULTS Vector containing results that summerize the TMD inferences
}
\description{
This function returns a vector with results from DeepTMHMM
}
\examples{

################
# load depends #
################
library(seqinr)

#########################
# read input fasta file #
#########################
handle.InputFastaFile <- Function.ReadORFasta(string.PathFastaFile = '~/Desktop/Beetle ORs Protein Translations.fasta')

##################################
## run Deep TMHMM on single gene #
##################################
#Function.RunDeepTMHMM(string.SequenceName = handle.InputFastaFile$SequenceName[1],
#                      string.SequenceData = handle.InputFastaFile$SequenceData[1], 
#                      string.SpeciesName = handle.InputFastaFile$SequenceSpecies[1], 
#                      string.SequenceScaffold = handle.InputFastaFile$SequenceScaffolds[1],
#                      string.PathDir = '~/Desktop/')

################
# read results #
################
Function.ReadResults_DeepTMHMM(string.SequenceName = handle.InputFastaFile$SequenceName[1],
                               string.SpeciesName = handle.InputFastaFile$SequenceSpecies[1], 
                               string.SequenceScaffold = handle.InputFastaFile$SequenceScaffolds[1],
                               string.Path_RESULTS = '~/Desktop/Conduct.DeepTMHMM_AglaOr1_Orco_2022-08-02/biolib_results/')
}
\keyword{beetle}
\keyword{chemosensation,}
\keyword{coleoptera,}
\keyword{genomes,}
\keyword{olfactory}
\keyword{receptors,}
