% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.RunDeepTMHMM.R
\name{Function.RunDeepTMHMM}
\alias{Function.RunDeepTMHMM}
\title{Function.RunDeepTMHMM: function to run DeepTMHMM on a single amino acid sequence}
\usage{
Function.RunDeepTMHMM(
  string.SequenceName,
  string.SpeciesName,
  string.SequenceScaffold,
  string.SequenceData,
  string.PathDir
)
}
\arguments{
\item{string.SequenceName}{String containing the name of the sequence (used to name the directory as well)}

\item{string.SpeciesName}{String containing the name of species (genome) taken from the original fasta file}

\item{string.SequenceScaffold}{String containing the name of the scaffold with putative BLAST hits}

\item{string.SequenceData}{String containing the sequence data for a single sequence to run DeepTMHMM}

\item{string.PathDir}{String containing the directory path to where the analyses will be conducted}
}
\value{
Directory External directory with results from DeepTMHMM
}
\description{
This function returns a subdirectory containing the results from DeepTMHMM
}
\examples{

################
# load depends #
################
library(seqinr)

#########################
# read input fasta file #
#########################
handle.InputFastaFile <- Function.ReadORFasta(string.PathFastaFile = '~/Desktop/ChemoGenes_ProteinSet1.fasta')

#################################
# run Deep TMHMM on single gene #
#################################
Function.RunDeepTMHMM(string.SequenceName = handle.InputFastaFile$SequenceName[1],
                      string.SequenceData = handle.InputFastaFile$SequenceData[2], 
                      string.PathDir = '~/Desktop/')
}
\keyword{beetle}
\keyword{chemosensation,}
\keyword{coleoptera,}
\keyword{genomes,}
\keyword{olfactory}
\keyword{receptors,}
