% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.RunDeepTMHMM_Pipeline.R
\name{Function.RunDeepTMHMM_Pipeline}
\alias{Function.RunDeepTMHMM_Pipeline}
\title{Function.RunDeepTMHMM_Pipeline: function to run DeepTMHMM on each sequence of a AA fast file}
\usage{
Function.RunDeepTMHMM_Pipeline(string.PathDir, string.FastaFile)
}
\arguments{
\item{string.PathDir}{String containing the directory path to where the analyses will be conducted}

\item{string.FastaFile}{String containing the path to the input AA fasta file}
}
\value{
handle.DF_Results Data.frame containing results from fasta file
}
\description{
This function returns a data.frame containing results of DeepTMHMM running on each sequence in fasta file
}
\examples{

################
# load depends #
################
library(seqinr)

##############################
# run on complete fasta file #
##############################
handle.RESULTS <- Function.RunDeepTMHMM_Pipeline(string.PathDir = '~/Desktop/', string.FastaFile = '~/Desktop/Test.fasta')

ANOTHER EXAMPLE:
################
# Load depends #
################
library(ColeOR)
library(seqinr)

#####################################
# Read example chromosome alignment #
#####################################
String.Path_ExampleModels <-  system.file("extdata", "ExampleModelsORs.fasta", package="ColeOR")

#######################
# run TMD predictions #
#######################
handle.RESULTS <- Function.RunDeepTMHMM_Pipeline(string.PathDir = '~/Desktop/', 
                                                string.FastaFile = String.Path_ExampleModels)
}
\keyword{beetle}
\keyword{chemosensation,}
\keyword{coleoptera,}
\keyword{genomes,}
\keyword{olfactory}
\keyword{receptors,}
