% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function.WriteORFasta.R
\name{Function.WriteORFasta}
\alias{Function.WriteORFasta}
\title{Function.WriteORFasta: function to create directory and write file for amino acid sequence for OR gene}
\usage{
Function.WriteORFasta(
  string.SpeciesName,
  string.SequenceName,
  string.SequenceScaffold,
  string.SequenceData,
  string.PathDir
)
}
\arguments{
\item{string.SpeciesName}{String containing the name of species (genome) taken from the original fasta file}

\item{string.SequenceName}{String containing the name of the OR sequence}

\item{string.SequenceScaffold}{String containing the name of the scaffold with putative BLAST hits}

\item{string.SequenceData}{String containing the sequence data (amino acid sequence from Function.ReadORFasta)}

\item{string.PathDir}{String containing the directory path to where the analyses will be conducted}
}
\value{
File Function creates a directory and writes OR amino acid sequence to file in directory
}
\description{
This function returns a directory with fasta file containing amino acid sequence
}
\examples{

################
# load depends #
################
library(seqinr)

#########################
# read input fasta file #
#########################
handle.InputFastaFile <- Function.ReadORFasta(string.PathFastaFile = '~/Desktop/Beetle ORs Protein Translations.fasta')

#################################
# run Deep TMHMM on single gene #
#################################
Function.WriteORFasta(string.SequenceName = handle.InputFastaFile$SequenceName[1],
                      string.SpeciesName = handle.InputFastaFile$SequenceSpecies[1],
                      string.SequenceScaffold = handle.InputFastaFile$SequenceScaffolds[1],
                      string.SequenceData = handle.InputFastaFile$SequenceData[1], 
                      string.PathDir = '~/Desktop/')
}
\keyword{beetle}
\keyword{chemosensation,}
\keyword{coleoptera,}
\keyword{genomes,}
\keyword{olfactory}
\keyword{receptors,}
